requirements.txt
distro==1.8.0
ghostscript==0.7
numpy==1.24.2
pandas==2.0.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabula-py==2.7.0
tzdata==2023.3
